requirements.txt
# Z3ube Backend Requirements

# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# AI/ML Libraries
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.0.2
langchain-ollama==0.1.0

# Vector embeddings and similarity
sentence-transformers==2.3.1
faiss-cpu==1.13.2

# Knowledge graphs and data structures
networkx==3.2.1
pyvis==0.3.2

# Data processing
numpy==1.26.3
pandas==2.2.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Caching (optional)
redis==5.0.1
hiredis==2.3.2

# Web scraping and research
beautifulsoup4==4.12.3
requests==2.31.0
httpx==0.26.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
